SCB-150 update package name and resolve conflicts
