+++ metrics/metrics-core/src/test/java/org/apache/servicecomb/metrics/core/TestEventAndRunner.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package io.servicecomb.metrics.core;
+package org.apache.servicecomb.metrics.core;
 
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.mockito.Mockito.when;
@@ -29,21 +29,20 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.servicecomb.core.metrics.InvocationFinishedEvent;
+import org.apache.servicecomb.core.metrics.InvocationStartProcessingEvent;
+import org.apache.servicecomb.core.metrics.InvocationStartedEvent;
+import org.apache.servicecomb.foundation.common.utils.EventUtils;
+import org.apache.servicecomb.metrics.common.MetricsDimension;
+import org.apache.servicecomb.metrics.common.RegistryMetric;
+import org.apache.servicecomb.metrics.core.event.DefaultEventListenerManager;
+import org.apache.servicecomb.metrics.core.monitor.DefaultSystemMonitor;
+import org.apache.servicecomb.metrics.core.monitor.RegistryMonitor;
+import org.apache.servicecomb.metrics.core.publish.DefaultDataSource;
+import org.apache.servicecomb.swagger.invocation.InvocationType;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
-
-import io.servicecomb.core.metrics.InvocationFinishedEvent;
-import io.servicecomb.core.metrics.InvocationStartProcessingEvent;
-import io.servicecomb.core.metrics.InvocationStartedEvent;
-import io.servicecomb.foundation.common.utils.EventUtils;
-import io.servicecomb.metrics.common.MetricsDimension;
-import io.servicecomb.metrics.common.RegistryMetric;
-import io.servicecomb.metrics.core.event.DefaultEventListenerManager;
-import io.servicecomb.metrics.core.monitor.DefaultSystemMonitor;
-import io.servicecomb.metrics.core.monitor.RegistryMonitor;
-import io.servicecomb.metrics.core.publish.DefaultDataSource;
-import io.servicecomb.swagger.invocation.InvocationType;
 
 public class TestEventAndRunner {
